Update whats_new version, move dtypes tests to new file and add tests
